+++ Assets/Utils/NormalMapGeneration/NormalMapSettings.cs
@@ -3,11 +3,11 @@ using UnityEngine;
 public struct NormalMapSettings
 {
     public Texture2D sourceTexture;
-    public float strengthEdges;
+    public int strengthEdges;
     public int blurEdgesRadius;
-    public float strengthBorder;
+    public int strengthBorder;
     public int blurBorderRadius;
-    public float softenBorder;
+    public int softenBorder;
     public float slopePercentageBorder;
     public int finalBlurRadius;
 }
